% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_survey_question_aux.R
\name{apply_labels}
\alias{apply_labels}
\title{Aplicação de rótulo}
\usage{
apply_labels(data, question_label, answer_label)
}
\arguments{
\item{data}{Tibble. Dados já organizados com os percentuais calculados}

\item{question_label}{Tibble. Dados gerados pelo usuário com os rótulos
de questão vinculado aos identificadores de questão}

\item{answer_label}{Tibble. Dados gerados pelo usuário com os rótulos
de respostas vinculadas aos identificadores de questão}
}
\description{
Utiliza os dados de rótulo de questões e resposta para aplicar aos dados
}
