% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_text.R
\name{break_text}
\alias{break_text}
\title{Break text}
\usage{
break_text(string, max_size)
}
\arguments{
\item{string}{Character. O texto deve ser quebrado em partes menores}

\item{max_size}{Numeric. Limite de tamanho por linha}
}
\value{
Character. O mesmo texto anterior, porém com quebras de linhas.
}
\description{
Quebra o texto por linhas segundo um tamanho máximo de caracter por linha.
}
\details{
Essa função insere quebras de texto "\\n" segundo um
tamanho máximo de caracteres por linha.
A quebra é inserida logo após a palavra que ultrapassa o limite por linha.
Utilizada para quebra de textos que vão para o gráfico de modo que possa ser
cursomizado de forma mais automática o tamanho horizontal de textos longos.
}
\author{
Leonardo Rocha
}
