% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_survey_question_aux.R
\name{extract_p}
\alias{extract_p}
\title{Extração de p-valor}
\usage{
extract_p(x)
}
\arguments{
\item{x}{Tibble. tabela de contingência}
}
\value{
Numeric. P-valor relativo à contingência
}
\description{
Extração de p-valor por um teste de aderência do tipo da qui-quadrado,
aplicado a uma tabela de contingência. Utilizado dentro da tabulação das
questões do survey para gerar a coluna de p-valor porquestão principal.
}
