% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_survey_question_aux.R
\name{gen_names}
\alias{gen_names}
\title{gerar nomes}
\usage{
gen_names(number)
}
\arguments{
\item{number}{Numeric. Gera o nome da questão a partir do número}
}
\value{
Character. Nome da questão principal composto pelo número precedido
pelo caracter "q".
}
\description{
gerar nomes
}
