% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_tabs.R
\name{link_tabs}
\alias{link_tabs}
\title{Relacionar tabelas}
\usage{
link_tabs(data, qfather, nson)
}
\arguments{
\item{data}{List. A lista com todos os dados nomeados pelo nome da questão}

\item{qfather}{Numeric. O número da questão pai}

\item{nson}{Numeric. O número da questão filha}
}
\value{
Tibble. O dado da filha com a coluna respectiva ao nome da pai
}
\description{
Relaciona duas tabelas de frequências de respostas de questões relacionadas no survey
}
\details{
Pega a tabela de uma questão que é considerada filha de outra, e seleciona apenas as contagens de
nível de conhecimento total, e cria uma nova coluna que é equivalente à opção equivalente da questão pai.
Essa função deve ser aplicada a todas as questões filhas de um mês pai para depois a lista de resultados
dessa função ser empilhada pela rowbindlist.
Para o bom funcionamento da questão a numeração das respostas da questão pai precisa seguir
a ordem da questão filha na lista de perguntas. Ou seja se, a opção 1 é equivalente à questão imediamente
após a questão pai.
}
\author{
Leonardo Rocha
}
