% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bar_graph.R
\name{make_bar_graph}
\alias{make_bar_graph}
\title{Geração de gráfico de barras}
\usage{
make_bar_graph(
  data,
  negative_col = FALSE,
  fill_color = "Oranges",
  title_max_size = 60,
  legend_max_size = 10,
  axis_max_size = 25,
  sort_by_value = FALSE,
  legend_position = "bottom",
  base_text_size = 9,
  title = FALSE,
  grey_scale = FALSE,
  with_labels = TRUE
)
}
\arguments{
\item{with_labels}{}
}
\value{
Gráfico customizado de barras na horizontal gerado pelo ggplot
}
\description{
longo
}
\details{
descrição
}
\author{
Leonardo Rocha
}
