% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_analysis.R
\docType{package}
\name{survey.analysis}
\alias{survey.analysis}
\alias{survey.analysis-package}
\title{survey.analyis: pacote para análise de dados do Survey Monkey}
\description{
Esse pacote fornece ferramentas para a realizações de tabulações,
testes de aderência e gráficos para dados de pesquisas realizados
a partir do Survey Monkey
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leofmr/survey.analysis}
  \item Report bugs at \url{https://github.com/leofmr/survey.analysis/issues}
}

}
\author{
Leonardo Rocha \email{leonardo.rocha@fiocruz.br}
}
