% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_survey_question.R
\name{tab_suvey_question}
\alias{tab_suvey_question}
\title{Tabulação de questão do Survey Monkey}
\usage{
tab_suvey_question(dt, question_number, question_label, answer_label)
}
\arguments{
\item{dt}{Tibble. dataset composto por todas as respostas do survey}

\item{question_number}{Numeric. número da questão que vai ser tabulada}

\item{answer_label}{Tibble. tabela com a relação entre questão principal e legendas de respostas.}

\item{question_labe}{Tibble. tabela com a relação de indice de questão e legendas de questão.}
}
\value{
Tibble com as frequências por nível de conhecimento, resposta e subquestões calculadas,
com as questões e opções rotuladas, e o p-valor do teste de aderência por nível de conhecimento
}
\description{
Realiza as tabulações para uma questão do survey monkey,
calculando o p-valor para testar aderência entre nível de conhecimento.
}
\details{
Essa função irá fazer as tabulações com percepção das escolhas para cada resposta, de questões principais.
Dado um número de questão principal, as respostas de todas as subquestão são também calculadas.
}
\author{
Leonardo Rocha
}
