% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_treemap_graph.R
\name{make_treemap_graph}
\alias{make_treemap_graph}
\title{Geração de gráfico treemap}
\usage{
make_treemap_graph(
  data,
  legend_max_size = 15,
  fill_color = "Set2",
  grey_scale = FALSE,
  with_labels = TRUE
)
}
\arguments{
\item{data}{Tibble. survey dataset completo}

\item{legend_max_size}{Numeric. Tamanho máximo da linha do texto legenda do gráfico}

\item{fill_color}{Character. Paleta de cores do preenchimento do treemap e da barra}

\item{grey_scale}{Logical. indicativo se deve ser utilizada escala de cinza}

\item{with_labels}{Logical. indicativo se desemv ser adicionados os percentuais no gráfico}
}
\value{
Gráfico customizado de barras na horizontal gerado pelo ggplot
}
\description{
longo
}
\details{
descrição
}
\author{
Leonardo Rocha
}
